rag_app/requirements.txt
streamlit==1.32.0

langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.2

chromadb==0.4.24

pypdf==4.2.0
docx2txt

sentence-transformers==2.6.1

transformers==4.40.1
huggingface_hub==0.23.0

python-dotenv==1.0.1
tqdm==4.66.4
